+++ src/util/Utility.hpp
@@ -18,7 +18,7 @@ namespace util {
 bool createTestFile(const std::string& fileName, uint64_t count, std::function<int32_t(int32_t)> factory);
 
 /// Read a file and invoke callback function on every entry
-bool foreachInFile(const std::string& fileName, std::function<void(uint64_t)> callback);
+bool foreachInFile(const std::string& fileName, std::function<void(uint32_t)> callback);
 
 /// Create a file with random(not set) data of given size
 bool createFile(const std::string& fileName, const uint64_t bytes);
